Fuse file tests, To/FromJSON on File, Sum methods
This starts towards making the fuse test useful, it just populates the
memory client with files so far.  It still needs better timing
coordination, and to check the fuse files.  Manual testing indicates
there's some other bug to fix, as well.

I hung some helper methods off of the File struct to marshal and
unmarshal them.  I'm not confident that this will be useful long term,
so I haven't converted the rest of the codebase to use it yet.

I am pretty confident that we want to centralize Sum, but possibly
through methods on Chunk and File?

I added a lot of comments to File and Chunk while in there.  The AES
documentation is a bit speculative, and I have an inkling that perhaps
a level of indirection will be necessary, eg.
Encrypted: {
  Method: "AESGCM"
  Key:    ""
}
... we'll see how that goes.