Move cache->tree, change drive.Cache.NewClient
This revamps the caching inside the FusFS to use the new drive.Cache
interface.  I changed the NewClient constructor to take a slice of
clients, as returned by config.Clients(), rather than a slice of
configs.  This seemed like a more natural interface for what to pass to
cache.Drive, but required a bit of fiddling with its tests.

This also drops the "automatic replication" behavior of the Fuse FS.  It
doesn't write back any chunk it sees to all clients.  I plan to write a
separate "branch" job to do that.  Branch, heh.  Get it?   It makes
shade off to the side.

^^ That is why I write subroutines instead of stand up routines...