Lock memory client correctly, add parallel tests
Fix a bug in the memory client where it wasn't locking the map of chunks
when adding to it.

Throw much more test data through the List/Put files test helper.

Add parallel test helpers, and invoke them in the memory and local
Client tests.

Reduce the number of times we checksum a random chunk of test data, and
avoid the need to hex encode/decode the strings over and over.