fix readDir bug, impl. read, add read tests
readDir depends on a map to store children, but the kernel calls readDir
multiple times at various offsets, expecting a stable response each
time.  Finding and fixing that was... tedious.

tests for reading from the fuse client, via files and chunks.  Added
code to allow refreshing from drive.Client on demand for those.

Disable parallel servicing of FUSE requests in multiple goroutines, as
the tests show that it's much slower to do it that way.  The real world
with latent disks and cloud backends might change that.