Improve parallel testing of MaxFiles
Testing in parallel requires each test to have control over the
relationship between the number of files which will be used in the round
rip test, and the MaxFiles supported by the client.  This extenalizes
the number of files, so they can be configured to test both LRU behavior
and sane parallel behavior.