Refactor drive.Config to be recursive
This changes the config from a []drive.Config, each of which described a
backend, to a single drive.Config object.  The drive/cache package is
now just-another-shade.Drive with a registered Provider type of "cache";
it supports the standard drive.NewClient interface.  The drive.Config
object now contains a "Children" member, which you can pass a
[]drive.Config to the "cache" Provider type to specify how to group more
than one backend.

This lays the groundwork for having nested caches, some of which are
encrypted, some of which are not.  It also centralizes the complex logic of
instantiating multiple clients, writing to them in parallel, etc. in the
drive/cache package, so that it can be shared between the various tools.