Use Google Drive AppData folder when parent IDs absent
If either the ChunkParentID or FileParentID options are not specified, the
files which would have been written to those directories is instead written to
the hidden app data directory. The result is an uncluttered Google Drive file
list. Data stored this way can be wholesale deleted from the "Settings" pane.