Cache now writes files back to local storage
Kind of key to the idea of the "cache" is that it, well, you know...
caches things.  :)  It now does that.

In contemplating how many workers to run, I also renamed the numWorks
flag to numFuseWorkers.  I settled on just 1 cache worker per type,
queueing requests in the buffered channel, but kept the flag rename as
we'll probably end up with some other type of worker in shade
eventually.