Fetch chunk again during chunk round trip; lint
Fetching the chunk again during TestChunkRoundTrip avoids any ambiguity
about if a Get or a Put updates the LRU status (the test now exercies
them both the same).  It might be a reasonable LRU behavior for
MaxChunkBytes to only preserve the most recently Put chunk, or Read
chunk, so the test avoids being sensitive to that choice.

I also renamed a few variables in the same test, and cleaned up the
method of testing which chunks are retained vs. expired.  This allows
test to be configured where all chunks will be retained.  Otherwise, the
parallel test was expecting some impossible-to-predict half the chunks
to be retained.