Gut and replace the memory cache with an LRU.
This replaces the bespoke memory cache with a thin wrapper around
hashicorp/golang-lru.  The memory client needed MaxCacheBytes and
MaxFiles behavior, and this seems much wiser than reimplementing an LRU.

The tests changed to match, and I added a few tests to cover edge cases
that I had to rediscover (eg. that callers can get tripped up if they
modify the returned chunks and that modifies the cache).

I also took the opportunity to move the comparison test code from the
cache client into the memory client's API, simplifying the cache
client's tests.