Add Release{File,Chunk} and ChunkLister interface
This lays the foundation for cleaning up unused chunks.  It adds support
for listing chunks in all of the Drive implementations, and (optional)
support for deleting files and chunks.

There is also minor refactoring in the drive/oogle package to enable
some code reuse, and save a bit of ram on the stack.