Make cache implementation more deterministic
The buffered channel and goroutines for refreshes were causing
non-deterministic behavior; when Gets queued lots of refreshes, they
could complete after a Release of the same chunk.  Since refreshes
already only write to Local() clients, I decided its better to be
deterministic and just eat the latency on Get().  I optimized the
refresh so it didn't write back to the same client and ripped out the
goroutines and channels.

This does mean we'll have to copy a chunk to memory and local disk
before it can be returned from remote storage, but we don't require a
Flush() to local disk so that's mostly just writing it to a disk buffer,
which should be acceptable latency.