Add Warm function, to prefetch chunks if necessary
Warm is a new function on the Drive interface which allows the caller to
indicate to the Drive client that it is likely to fetch the indicated
chunks in the near future.

This is used by the Google Drive client to batch up metadata requests,
reducing the necessary number of API calls, and avoiding the latency hit
of an extra API round trip before each chunk is downloaded.